cleanup of visual studio compiler warnings
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1818 d60aafaf-7936-0410-ba4d-d307febf7868

+++ src/Console.cpp
@@ -494,7 +494,7 @@ ConsoleWindow::ConsoleWindow ()
   cons->buffer->setCaretVisible(false);
 
   cons->buffer->setFont(Font(Font::getDefaultMonospacedFontName(),
-			     Preferences::getInstance()->getIntProp(T("ConsoleFontSize"), 16),
+			     (float)Preferences::getInstance()->getIntProp(T("ConsoleFontSize"), 16),
 			     Font::plain));
   //std::cout << Font::getDefaultMonospacedFontName().toUTF8() << "\n";
